+++ fxmanifest.lua
@@ -14,7 +14,7 @@ client_scripts {
 }
 
 server_scripts {
-    'server.lua',
+    'server.js',
 }
 
 files {