+++ html/js/style.js
@@ -1,7 +1,11 @@
-import { style } from "../config.js";
+import { getHandoverData } from "./util/handover.js";
+
+const {
+    config: { style },
+} = getHandoverData();
 
 const link = document.createElement("link");
-link.href = style;
+link.href = `./styles/${style}.css`;
 link.rel = "stylesheet";
 
 document.getElementsByTagName("head")[0].appendChild(link);